RAND: cleanup after merge

+++ src/rand.h
@@ -22,6 +22,7 @@
 #ifndef NTRU_RAND_H
 #define NTRU_RAND_H
 
+
 #include "context.h"
 
 #include <tompoly.h>
@@ -36,8 +37,12 @@
  */
 #define GET_INT_FROM_RRAND 3
 
-pb_poly *ntru_get_random_poly_ternary(int length, int entropy_source);
-pb_poly *ntru_get_random_poly(size_t length, int entropy_source, mp_int *upper_bound,
+
+pb_poly *ntru_get_random_poly_ternary(size_t length, int entropy_source);
+pb_poly *ntru_get_random_poly(int length,
+		int entropy_source,
+		mp_int *upper_bound,
 		mp_int *lower_bound);
+
 
 #endif /* NTRU_RAND_H */
